% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{searchReads}
\alias{searchReads}
\title{Get reads from a GA4GH compliant server.}
\usage{
searchReads(base_url,
  readGroupIds = list("MWtnLXAzLXN1YnNldDpIRzAwNTM0OkVSUjAyMDIzOA=="),
  referenceId = "R1JDaDM4LXN1YnNldDoz", start = NULL, end = NULL,
  fields = NULL, converter = c)
}
\arguments{
\item{base_url}{the base URL of the GA4GH server to which
requests will be sent to.}

\item{readGroupIds}{The read group set ID.}

\item{referenceId}{The chromosome.}

\item{start}{Start position on the chromosome in 0-based coordinates.}

\item{end}{End position on the chromosome in 0-based coordinates.}

\item{fields}{A subset of fields to retrieve.  The default (NULL) will
return all fields.}

\item{converter}{A function that takes a list of read R objects and returns
them converted to the desired type.}
}
\value{
By default, the return value is a list of R objects
corresponding to the JSON objects returned by the GA4GH
Reads API.  If a converter is passed, object(s) of the type
returned by the converter will be returned by this function.
}
\description{
This function will return all of the reads that comprise the requested
genomic range, iterating over paginated results as necessary.
}
\details{
Optionally pass a converter as appropriate for your use case.  By passing it
to this method, only the converted objects will be accumulated in memory. The
converter function should return an empty container of the desired type
if called without any arguments.
}

