% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{searchReadsPage}
\alias{searchReadsPage}
\title{Get one page of reads from a GA4GH compliant server.}
\usage{
searchReadsPage(base_url,
  readGroupIds = list("MWtnLXAzLXN1YnNldDpIRzAwNTM0OkVSUjAyMDIzOA=="),
  referenceId = "R1JDaDM4LXN1YnNldDoz", start = NULL, end = NULL,
  fields = NULL, pageToken = NULL)
}
\arguments{
\item{base_url}{the base URL of the GA4GH server to which
requests will be sent to.}

\item{readGroupIds}{The read group set ID.}

\item{referenceId}{The chromosome.}

\item{start}{Start position on the chromosome in 0-based coordinates.}

\item{end}{End position on the chromosome in 0-based coordinates.}

\item{fields}{A subset of fields to retrieve.  The default (NULL) will
return all fields.}

\item{pageToken}{The page token. This can be NULL (default) for the first
page.}
}
\value{
A two-element list is returned by the function.

    reads: A list of R objects corresponding to the JSON objects returned
              by the GA4GH Reads API.

    nextPageToken: The token to be used to retrieve the next page of
                   results, if applicable.
}
\description{
In general, use the searchReads method instead.  It calls this method,
returning reads from all of the pages that comprise the requested
genomic range.
}
\details{
Note that the Global Alliance for Genomics and Health API uses a 0-based
coordinate system.  For more detail, please see GA4GH discussions such
as the following:
\itemize{
   \item\url{https://github.com/ga4gh/schemas/issues/168}
   \item\url{https://github.com/ga4gh/schemas/issues/121}
}
}
\examples{
readsPage <- searchReadsPage()
summary(readsPage)
summary(readsPage$reads[[1]])
}
\seealso{
Other page fetch functions: \code{\link{searchVariantsPage}}
}

