% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{searchVariantsPage}
\alias{searchVariantsPage}
\title{Get one page of variants from a GA4GH compliant server.}
\usage{
searchVariantsPage(base_url, variantSetId = "MWtnLXAzLXN1YnNldDptdm5jYWxs",
  referenceName = "1", start = NULL, end = NULL, fields = NULL,
  pageToken = NULL)
}
\arguments{
\item{base_url}{the base URL of the GA4GH server to which
requests will be sent to.}

\item{variantSetId}{The variant set ID.}

\item{referenceName}{The chromosome.}

\item{start}{Start position on the chromosome in 0-based coordinates.}

\item{end}{End position on the chromosome in 0-based coordinates.}

\item{fields}{A subset of fields to retrieve.  The default (NULL) will
return all fields.}

\item{pageToken}{The page token. This can be NULL (default) for the first
page.}
}
\value{
A two-element list is returned by the function.

    variants: A list of R objects corresponding to the JSON objects returned
              by the GA4GH Variants API.

    nextPageToken: The token to be used to retrieve the next page of
                   results, if applicable.
}
\description{
In general, use the getVariants method instead.  It calls this method,
returning variants from all of the pages that comprise the requested
genomic range.
}
\examples{
variantsPage <- getVariantsPage()
summary(variantsPage)
summary(variantsPage$variants[[1]])
}
\seealso{
Other page fetch functions: \code{\link{searchReadsPage}}
}

