% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{sendRequest}
\alias{sendRequest}
\title{Get one page of search results for a particular endpoint from
a GA4GH compliant server.}
\usage{
sendRequest(base_url, endpoint, method, body, query_fields, pageToken)
}
\arguments{
\item{base_url}{the base URL of the GA4GH server to which
requests will be sent to.}

\item{endpoint}{a endpoint such as reads, variants,
variantSets, etc...}

\item{method}{which method (GET/POST) should be used in the request to
the API?}

\item{body}{The body of the message to GET/POST to the endpoint.}

\item{query_fields}{The fields to be returned in the search response.}

\item{pageToken}{The page token. This can be NULL for the first page.}
}
\value{
The raw response converted from JSON to an R object.

}
\description{
In general, use higher level methods such as getReads and getVariants
instead.
}
\examples{
body <- list(readGroupSetIds=list("CMvnhpKTFhDnk4_9zcKO3_YB"),
             referenceName="22",
             start=16051400, end=16051500, pageToken=NULL)
reads <- getSearchPage("reads", body, NULL, NULL)
summary(reads)
}

