% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variantsToVRanges}
\alias{variantsToVRanges}
\title{Convert variants to VRanges.}
\usage{
variantsToVRanges(variants, oneBasedCoord = TRUE, slStyle = "UCSC")
}
\arguments{
\item{variants}{A list of R objects corresponding to the JSON objects
returned by the GA4GH Variants API.}

\item{oneBasedCoord}{Convert genomic positions to 1-based coordinates.}

\item{slStyle}{The style for seqnames (chrN or N or...).  Default is UCSC.}
}
\value{
\link[VariantAnnotation]{VRanges}
}
\description{
Note that the Global Alliance for Genomics and Health API uses a 0-based
coordinate system.  For more detail, please see GA4GH discussions such
as the following:
\itemize{
   \item\url{https://github.com/ga4gh/schemas/issues/168}
   \item\url{https://github.com/ga4gh/schemas/issues/121}
}
}
\examples{
variants1 <- searchVariants(converter = variantsToVRanges)
summary(variants1)
variants2 <- variantsToVRanges(searchVariants())
print(identical(variants1, variants2))
}
\seealso{
Other variants converter functions: \code{\link{variantsToGRanges}}
}

